Fix integration tests
Migrate logic to top-level omero_cli_metadata and populate_metadata packages to
prevent shadowing.
